Update Duplicates merge functions and table

+++ app/manage_cars.php
@@ -17,7 +17,7 @@ if (!securePage($_SERVER['PHP_SELF'])) {
 $badusers="SELECT users.id FROM users  LEFT JOIN car_user ON (users.id = car_user.userid) WHERE ( users.email_verified = 0 AND users.last_login = 0 AND car_user.carid IS NULL AND vericode_expiry < CURRENT_DATE ) GROUP BY users.id ";
 $unusedprofiles=" SELECT t1.user_id FROM profiles t1 LEFT JOIN users t2 ON t1.user_id = t2.id WHERE t2.id IS NULL ";
 $orphanedcars="SELECT t1.userid FROM car_user t1 LEFT JOIN users t2 ON t1.userid = t2.id  WHERE t2.id IS NULL ";
-$duplicates="SELECT a.* FROM cars a JOIN( SELECT chassis, COUNT( * ) FROM users_carsview WHERE chassis <> '' GROUP BY chassis HAVING COUNT( * ) > 1 ) b ON a.chassis = b.chassis ORDER BY a.chassis DESC";
+$duplicates="SELECT  a.* FROM cars a JOIN(  SELECT  type,chassis,  COUNT(*)  FROM  users_carsview  WHERE chassis <> '' GROUP BY type,chassis HAVING COUNT(*) > 1) b ON a.chassis = b.chassis ORDER BY a.chassis, a.type";
 
 // Get list of suspected duplicates
 $duplicatesQ = $db->query($duplicates);
@@ -78,31 +78,41 @@ if (!empty($_POST)) {
                         break;
 
                     case "merge":
-                        $new_car_id = $_POST['new_car_id'];
-                        $old_car_id  = $_POST['old_car_id'];
+                            $cars = $_POST['cars'];
+                            $N = count($cars);
+                            if ($N <> 2) {
+                                $errors[] = 'Select 2 cars to merge';
+                            } else {
+                                if ($cars[0] > $cars[1]) {
+                                    $new_car_id = $cars[0];
+                                    $old_car_id = $cars[1];
+                                } else {
+                                    $new_car_id = $cars[1];
+                                    $old_car_id = $cars[0];
+                                }
+                                
+                                // Merge the history
+                                $db->query("UPDATE cars_hist SET car_id = ? WHERE car_id = ?", [$new_car_id, $old_car_id]);
+                                if ($db->error()) {
+                                    $errors[] = $db->errorString();
+                                    logger($user->data()->id, "User", "FAILED: Merged CAR $old_car_id to CAR $new_car_id.");
+                                } else {
+                                    // Unassign from the previous owner
+                                    $db->query("DELETE FROM car_user WHERE carid = ?", [$old_car_id]);
 
-                        // Merge the history
-                        $db->query("UPDATE cars_hist SET car_id = ? WHERE car_id = ?", [$new_car_id, $old_car_id]);
-                        if ($db->error()) {
-                            $errors[] = $db->errorString();
-                            logger($user->data()->id, "User", "FAILED: Merged CAR $old_car_id to CAR $new_car_id.");
-                        } else {
-                            // Unassign from the previous owner
-                            $db->query("DELETE FROM car_user WHERE carid = ?", [$old_car_id]);
+                                    // Remove old car
+                                    $db->query("DELETE FROM cars WHERE id = ?", [$old_car_id]);
 
-                            // Remove old car
-                            $db->query("DELETE FROM cars WHERE id = ?", [$old_car_id]);
+                                    // Add a record to the history with some information on the assignment
+                                    $fields['comments'] = "Car $old_car_id was merged with $new_car_id.  Car $old_car_id has been deleted.  This was most likely because of a duplicate entry.";
+                                    $fields['car_id'] = $new_car_id;
+                                    $fields['operation'] = "MERGE";
+                                    $db->insert("cars_hist", $fields);
 
-                            // Add a record to the history with some information on the assignment
-                            $fields['comments'] = "Car $old_car_id was merged with $new_car_id.  Car $old_car_id has been deleted.  This was most likely because of a duplicate entry.";
-                            $fields['car_id'] = $new_car_id;
-                            $fields['operation'] = "MERGE";
-                            $db->insert("cars_hist", $fields);
-
-                            $successes[] = 'Admin '.($user->data()->id).' Merged CAR '.$old_car_id.' to CAR '.$new_car_id;
-                            logger($user->data()->id, "User", "Merged CAR $old_car_id to CAR $new_car_id.");
-                        }
-                        
+                                    $successes[] = 'Admin '.($user->data()->id).' Merged CAR '.$old_car_id.' to CAR '.$new_car_id;
+                                    logger($user->data()->id, "User", "Merged CAR $old_car_id to CAR $new_car_id.");
+                                }
+                            }
                         // Now update the table
                         // Get list of suspected duplicates
                         $duplicatesQ = $db->query($duplicates);
@@ -138,7 +148,7 @@ if (!empty($_POST)) {
 				<div class="card card-default">
 				<div class="card-header"><h2><strong>Reassign Car</strong></h2></div>
 					<div class="card-body">
-					<form name="assignCar" action="manage_cars.php" method="POST" enctype="multipart/form-data">
+					<form name="assignCar" action="manage_car.php" method="POST" enctype="multipart/form-data">
 						<label for="car_id">Car ID:</label><br>
 						<input type="text" id="car_id" name="car_id"><br>
 						<label for="user_id">User ID:</label><br>
@@ -151,28 +161,7 @@ if (!empty($_POST)) {
 					</div> <!-- card-body -->
 				</div> <!-- card -->
 			</div> <!-- col -->
-
-
 			<div class="col-4" align="center">
-			<div class="card card-default">
-				<div class="card-header"><h2><strong>Merge Car</strong></h2></div>
-					<div class="card-body">
-					<form name="mergeCar" action="manage_cars.php" method="POST" enctype="multipart/form-data">
-						<label for="car_id">Older Car ID:</label><br>
-						<input type="text" id="old_car_id" name="old_car_id"><br>
-						<label for="user_id">New Car ID:</label><br>
-						<input type="text" id="new_car_id" name="new_car_id">
-						<br><br>
-						<input type="hidden" name="csrf" value="<?= Token::generate(); ?>" />
-						<input type="hidden" name="command" value="merge" />
-						<input class="btn btn-primary btn-lg btn-block" type='submit' value='Merge' class='submit' />
-					</form>
-					</div> <!-- card-body -->
-				</div> <!-- card -->
-			</div> <!-- col -->
-		</div> <!-- row -->
-		<div class="row">
-			<div class="col-12" align="center">
 				<div class="card card-default">
 				<div class="card-header"><h2><strong>Messages</strong></h2></div>
 					<div class="card-body">
@@ -190,71 +179,85 @@ if (!empty($_POST)) {
 		</div> <!-- row -->
 		<div class="row">
 			<div class="col" align="center">
-				<div class="card border-success">
-				<div class="card-header"><h2><strong>Suspected Duplicates?</strong></h2></div>
+				<div class="card card-default">
+				<div class="card-header"><h2><strong>Suspected Duplicates</strong></h2></div>
 					<div class="card-body">
-						<table id="duptable" class="table table-striped table-bordered table-sm" cellspacing="0" width="100%">
-							<thead>
-							<tr>
-								<th>CarID</th>
-								<th>Username</th>
-								<th>Create</th>
-								<th>Modified</th>
-								<th>Year</th>
-								<th>Type</th>
-								<th>Chassis</th>
-								<th>Series</th>
-								<th>Variant</th>
-								<th>Color</th>
-								<th>Engine</th>
-								<th>Purchase Date</th>
-								<th>Sold Date</th>
-								<th>Comments</th>
-								<th>Image</th>
-								<th>Fname</th>
-								<th>Lname</th>
-								<th>email</th>
-								<th>City</th>
-								<th>State</th>
-								<th>Country</th>
-							</tr>
-							</thead>
-							<tbody>
-							<?php
-                            //Cycle through users
-                            foreach ($duplicateCars as $v1) {
-                                ?>
-								<tr>
-								<td><a class="btn btn-success btn-sm" target="_blank" href='<?=$us_url_root?>app/car_details.php?car_id=<?=$v1->id?>'><?=$v1->id?></a></td>
+                        <form action="manage_cars.php" method="POST" enctype="multipart/form-data">
+                            <input type="hidden" name="command" value="merge" />
+                            <input type="hidden" name="csrf" value="<?= Token::generate(); ?>" />
 
-								<td><?=$v1->username?></td>
-								<td><?=$v1->ctime?></td> 
-								<td><?=$v1->mtime?></td> 
-								<td><?=$v1->year?></td>
-								<td><?=$v1->type?></td>
-								<td><?=$v1->chassis?></td>
-								<td><?=$v1->series?></td>
-								<td><?=$v1->variant?></td>
-								<td><?=$v1->color?></td>
-								<td><?=$v1->engine?></td>                        
-								<td><?=$v1->purchasedate?></td>
-								<td><?=$v1->solddate?></td>
-								<td><?=$v1->comments?></td>
-								<td> <?php
-                                if ($v1->image and file_exists($abs_us_root.$us_url_root."app/userimages/".$v1->image)) {
-                                    echo '<img src='.$us_url_root.'app/userimages/thumbs/'.$v1->image.">";
-                                } ?>  </td>
-								<td><?=$v1->fname?></td>
-								<td><?=$v1->lname?></td>
-								<td><?=$v1->email?></td>
-								<td><?=$v1->city?></td>
-								<td><?=$v1->state?></td>
-								<td><?=$v1->country?></td> 
-								</tr>
-							<?php
-                            } ?>
-							</tbody>
-						</table>
+                            <input type="submit" class="btn btn-primary btn-lg btn-block"  name="formSubmit" value="Merge" />
+                            <table id="duptable" class="table table-striped table-bordered table-sm" cellspacing="0" width="100%">
+                                <thead>
+                                <tr>
+                                    <th>Merge</th>
+                                    <th>CarID</th>
+                                    <th>Username</th>
+                                    <th>Create</th>
+                                    <th>Modified</th>
+                                    <th>Year</th>
+                                    <th>Type</th>
+                                    <th>Chassis</th>
+                                    <th>Series</th>
+                                    <th>Variant</th>
+                                    <th>Color</th>
+                                    <th>Engine</th>
+                                    <th>Purchase Date</th>
+                                    <th>Sold Date</th>
+                                    <th>Comments</th>
+                                    <th>Image</th>
+                                    <th>Fname</th>
+                                    <th>Lname</th>
+                                    <th>email</th>
+                                    <th>City</th>
+                                    <th>State</th>
+                                    <th>Country</th>
+                                </tr>
+                                </thead>
+                                <tbody>
+                                <?php
+                                //Cycle through users
+                                foreach ($duplicateCars as $v1) {
+                                    ?>
+                                    <tr>
+                                    <td class="center">
+                                        <div class="form-group">
+                                            <div class="custom-control custom-checkbox">
+                                            <input type="checkbox" class="custom-control-input" id="customCheck-<?=$v1->id?>" name="cars[]" value="<?=$v1->id?>" >
+                                            <label class="custom-control-label" for="customCheck-<?=$v1->id?>"></label>
+                                            </div>
+                                        </div>
+                                    </td>
+                                    <td><a class="btn btn-success btn-sm" target="_blank" href='<?=$us_url_root?>app/car_details.php?car_id=<?=$v1->id?>'><?=$v1->id?></a></td>
+                                    <td><?=$v1->username?></td>
+                                    <td><?=$v1->ctime?></td> 
+                                    <td><?=$v1->mtime?></td> 
+                                    <td><?=$v1->year?></td>
+                                    <td><?=$v1->type?></td>
+                                    <td><?=$v1->chassis?></td>
+                                    <td><?=$v1->series?></td>
+                                    <td><?=$v1->variant?></td>
+                                    <td><?=$v1->color?></td>
+                                    <td><?=$v1->engine?></td>                        
+                                    <td><?=$v1->purchasedate?></td>
+                                    <td><?=$v1->solddate?></td>
+                                    <td><?=$v1->comments?></td>
+                                    <td> <?php
+                                    if ($v1->image and file_exists($abs_us_root.$us_url_root."app/userimages/".$v1->image)) {
+                                        echo '<img src='.$us_url_root.'app/userimages/thumbs/'.$v1->image.">";
+                                    } ?>  </td>
+                                    <td><?=$v1->fname?></td>
+                                    <td><?=$v1->lname?></td>
+                                    <td><?=$v1->email?></td>
+                                    <td><?=$v1->city?></td>
+                                    <td><?=$v1->state?></td>
+                                    <td><?=$v1->country?></td> 
+                                    </tr>
+                                <?php
+                                } ?>
+                                </tbody>
+                            </table>
+                        </form>
 					</div> <!-- card-body -->
 				</div> <!-- card -->  
 			</div> <!-- col -->
@@ -380,13 +383,24 @@ if (!empty($_POST)) {
 <!-- Javascript -->
 
 <link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.18/css/jquery.dataTables.min.css"/>
+<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/fixedheader/3.1.2/css/fixedHeader.dataTables.min.css">
 <script type="text/javascript" src="https://cdn.datatables.net/1.10.18/js/jquery.dataTables.min.js"></script>
+<script type="text/javascript" src="https://cdn.datatables.net/rowgroup/1.1.2/js/dataTables.rowGroup.min.js"></script>
+<script src="https://cdn.datatables.net/fixedheader/3.1.2/js/dataTables.fixedHeader.min.js" type="text/javascript"></script>
+
+
 <script type="text/javascript">
 $(document).ready(function()  {
-var table =  $('#duptable').DataTable(
+$('#duptable').DataTable(
 	{
-	"ordering": false,
-	"scrollX": true
+    fixedHeader: true,
+	rowGroup: {
+            dataSrc: [6,7],    
+            startClassName: 'table-info',
+    },
+    "ordering": false,
+    "scrollX": true,
+    "pageLength": 50
 	});
 } );
 </script>
